backend/requirements.txt
pydantic
python-dotenv
mysql-connector-python
